+++ sdk/storage/azblob/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugs Fixed
 
 * `GetSASURL()`: for container and blob clients, don't add a forward slash before the query string
+* Fixed issue [#19249](https://github.com/Azure/azure-sdk-for-go/issues/19249) by increasing service version to '2020-02-10'.
 
 ### Other Changes